refactor(file_mounter): move FileMounter into new submodule
The FileMounter is just a light wrapper around the nix::mount call. It is moved to a separate submodule to keep things organized and modular.
